% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ardl-package.R
\docType{package}
\name{ARDL-package}
\alias{ARDL-package}
\title{ARDL: ARDL, ECM and Bounds-Test for Cointegration}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Creates complex autoregressive distributed lag (ARDL) models and constructs the underlying unrestricted and restricted error correction model (ECM) automatically, just by providing the order. It also performs the bounds-test for cointegration as described in Pesaran et al. (2001) \doi{10.1002/jae.616} and provides the multipliers and the cointegrating equation. The validity and the accuracy of this package have been verified by successfully replicating the results of Pesaran et al. (2001) in Natsiopoulos and Tzeremes (2022) \doi{10.1002/jae.2919}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Natsiopoulos/ARDL}
  \item Report bugs at \url{https://github.com/Natsiopoulos/ARDL/issues}
}

}
\author{
\strong{Maintainer}: Kleanthis Natsiopoulos \email{klnatsio@gmail.com} (\href{https://orcid.org/0000-0003-1180-2984}{ORCID})

Authors:
\itemize{
  \item Kleanthis Natsiopoulos \email{klnatsio@gmail.com} (\href{https://orcid.org/0000-0003-1180-2984}{ORCID})
  \item Daniel Finnan \email{dan@custom-made.org.uk} (\href{https://orcid.org/0009-0009-7468-5288}{ORCID})
}

Other contributors:
\itemize{
  \item Nickolaos Tzeremes \email{bus9nt@econ.uth.gr} (\href{https://orcid.org/0000-0002-6938-3404}{ORCID}) [thesis advisor]
}

}
\keyword{internal}
