% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.ardl.R
\name{auto.ardl}
\alias{auto.ardl}
\title{auto.ardl}
\usage{
auto.ardl(formula, data, subset = NULL, ymax = 4, xmax = 4, case = 3,
  ic = "bic", verbose = FALSE)
}
\arguments{
\item{formula}{Formula as in \code{ y ~ x1 + x2 | x3 } where \code{x1} and \code{x2} are tested for different lags up to \code{xmax}. Note that \code{x3} is "fixed" so it is not lagged.}

\item{data}{A dataframe or time referenced object with data in columns.}

\item{subset}{(optional) Filter rows from the dataframe. Defaults to \code{NULL}.}

\item{ymax}{(optional) Maximum lag of the dependent variable \code{y}. Defaults to 4.}

\item{xmax}{(optional) Maximum lag of the explaining variable(s) \code{x}. Defaults to 4.}

\item{case}{(optional) Number of the case according to PSS (2001) to chose which table to use.}

\item{ic}{(optional) Information criteria used for selection. Can be \code{"aic","bic","r2","ll"}. Defaults to "bic".}

\item{verbose}{(optional) Lists all the models tested. Defaults to FALSE.}
}
\value{
An object of class \code{ardl}.
}
\description{
Searchs for the best ARDL model, given maximum lag for the explained variable \code{y} and regressors \code{x}.
}
\details{
This function selects the models to test, the actual estimation is done by \code{ardl()}.
}
\examples{

data(br_month)

# show the selection process with verbose=TRUE
taylor0 <- auto.ardl( mpr~cpi+prod+reer, data=br_month, ymax=4, xmax=c(4,4,4), verbose=TRUE )
summary( taylor0 )

# note the differences by varying the case (intercept and trend) specification 
phillips1 <- auto.ardl( cpi~mpr+prod+reer|d_lula, data=br_month, ymax=4, xmax=c(4,4,4), case=1 )
phillips2 <- auto.ardl( cpi~mpr+prod+reer|d_lula, data=br_month, ymax=4, xmax=c(4,4,4), case=3 )
phillips3 <- auto.ardl( cpi~mpr+prod+reer|d_lula, data=br_month, ymax=4, xmax=c(4,4,4), case=5 )
BIC(phillips1) # 86.42611 best!
BIC(phillips2) # 91.0541
BIC(phillips3) # 92.84055

# change the information criteria 
phillips4 <- auto.ardl( cpi~mpr+prod+reer|d_lula, data=br_month, ymax=4, xmax=c(4,4,4), ic="aic" )
phillips5 <- auto.ardl( cpi~mpr+prod+reer|d_lula, data=br_month, ymax=4, xmax=c(4,4,4), ic="r2" )
phillips6 <- auto.ardl( cpi~mpr+prod+reer|d_lula, data=br_month, ymax=4, xmax=c(4,4,4), ic="bic" )
}
\seealso{
\code{\link{ardl}}
}

