% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.test.R
\name{bounds.test}
\alias{bounds.test}
\title{Performs PSS (2001) bounds test for I(0) and I(1) regressors.}
\usage{
bounds.test(obj, tables = "pss")
}
\arguments{
\item{obj}{An ardl object to be tested.}

\item{tables}{(optional) Select the tables with critical values, can be \code{"pss"} or \code{"narayan"}. Defaults to \code{"pss"}}
}
\value{
An object of class \code{ardl}.
}
\description{
Calculates the Pesaran, Shin and Smith (2001) statistics and list the critical values.
}
\details{
Presents the bounds test for I(0) or I(1) regressors with critical values from PSS (2001) or Narayan (2005) paper for small samples.
}
\examples{
data(br_month)
m1 <- ardl( mpr~cpi+prod+ner|d_lula, data=br_month, ylag=2, xlag=c(0,0,0), case=3 )
bounds.test( m1 )

}
\references{
Pesaran, M.H. and Shin, Yongcheol and Smith, Richard (2001) Bounds testing approaches to the analysis of level relationships. \emph{Journal of Applied Econometrics}.
}
\seealso{
\code{\link{coint}}
}

