% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_builders.R
\name{build_ardl_formula}
\alias{build_ardl_formula}
\title{ARDL formula specification builder}
\usage{
build_ardl_formula(parsed_formula, order)
}
\arguments{
\item{parsed_formula}{A list containing the formula parts as returned from
\code{\link{parse_formula}}.}

\item{order}{A numeric vector with the ARDL order as returned from
\code{\link[=parse_order]{parse_order(restriction = FALSE)}}.}
}
\value{
\code{build_ardl_formula} returns a list containing the full formula
  and the independent and dependent parts of the formula separated. The full
  formula is ready to be used as input in the \code{dynlm} function.
}
\description{
It creates the ARDL specification according to the given "formula" and their
corresponding "orders".
}
\seealso{
\code{\link{build_uecm_formula}}, \code{\link{build_recm_formula}}
}
\author{
Kleanthis Natsiopoulos, \email{klnatsio@gmail.com}
}
\keyword{internal}
