% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coint.R
\name{coint}
\alias{coint}
\title{Long and Short-Run coefficients}
\usage{
coint(obj, type = "txt", file = NULL, caption = NULL)
}
\arguments{
\item{obj}{An object of class \code{ardl}.}

\item{type}{(optional) Select results format that can be \code{c("txt","tex")}. Defaults to \code{"tex"}.}

\item{file}{(optional) Filename to save results. Defaults to \code{NULL}.}

\item{caption}{(optional) Caption (title) of the table. Defaults to \code{NULL}.}
}
\value{
Character strings.
}
\description{
Present LR and SR coefficients of an ardl (Auto Regressive Distributed Lag) object
}
\details{
Uses delta method to calculate standard-errors of the Long Run (LR) coefficients.
}
\examples{

data("br_month")
phillips <- ardl( cpi~prod+ner|d_lula, data=br_month, ylag=2, xlag=c(1,2), case=5 )
coint( phillips )
coint( phillips, type="txt", file="phillips.txt" )
coint( phillips, type="tex", file="phillips.tex" )
}
\seealso{
\code{\link{bounds.test}}
}

