% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{denmark}
\alias{denmark}
\title{The Danish data on money income prices and interest rates}
\format{
A time-series object with 55 rows and 5 variables. Time period from 1974:Q1
until 1987:Q3.
\describe{
  \item{LRM}{logarithm of real money, M2}
  \item{LRY}{logarithm of real income}
  \item{LPY}{logarithm of price deflator}
  \item{IBO}{bond rate}
  \item{IDE}{bank deposit rate}
}
}
\source{
\url{https://onlinelibrary.wiley.com/doi/10.1111/j.1468-0084.1990.mp52002003.x}
}
\usage{
denmark
}
\description{
This data set contains the series used by S. Johansen and K. Juselius for
estimating a money demand function of Denmark.
}
\details{
An object of class "zooreg" "zoo".
}
\references{
Johansen, S. and Juselius, K. (1990), Maximum Likelihood
Estimation and Inference on Cointegration -- with Applications to the Demand
for Money, \emph{Oxford Bulletin of Economics and Statistics}, \bold{52, 2},
169--210.
}
\keyword{datasets}
