% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ardl.R
\name{print.ardl}
\alias{print.ardl}
\title{Print ARDL model}
\usage{
\method{print}{ardl}(x, ...)
}
\arguments{
\item{x}{A class \code{ardl} object.}

\item{...}{ignored}
}
\description{
Generics to print the S3 class ardl (Auto Regressive Distributed Lag Model).
}
\details{
Calls \code{print.dynlm()}.
}

