% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ardl-package.R
\docType{package}
\name{ARDL-package}
\alias{ARDL}
\alias{ARDL-package}
\title{ARDL: ARDL, ECM and Bounds-Test for Cointegration}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Creates complex Autoregressive Distributed Lag models providing
    just the order and automatically constructs the underlying unrestricted and
    restricted ECM. It also performs the bounds-test for cointegration and
    provides the multipliers and the cointegrating equation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Natsiopoulos}
}

}
\author{
\strong{Maintainer}: Kleanthis Natsiopoulos \email{klnatsio@gmail.com} (\href{https://orcid.org/0000-0003-1180-2984}{ORCID}) [dissertant]

Other contributors:
\itemize{
  \item Nickolaos Tzeremes \email{bus9nt@econ.uth.gr} (\href{https://orcid.org/0000-0002-6938-3404}{ORCID}) [thesis advisor]
  \item University of Thessaly, Department of Economics [degree granting institution]
}

}
\keyword{internal}
