% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ardl-package.R
\docType{package}
\name{ARDL-package}
\alias{ARDL}
\alias{ARDL-package}
\title{ARDL: ARDL, ECM and Bounds-Test for Cointegration}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Creates complex autoregressive distributed lag (ARDL) models
    providing just the order and automatically constructs the underlying
    unrestricted and restricted error correction model (ECM). It also performs
    the bounds-test for cointegration as described in Pesaran et al. (2001) <doi:10.1002/jae.616> and provides the multipliers and the cointegrating
    equation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Natsiopoulos/ARDL}
}

}
\author{
\strong{Maintainer}: Kleanthis Natsiopoulos \email{klnatsio@gmail.com} (\href{https://orcid.org/0000-0003-1180-2984}{ORCID})

Authors:
\itemize{
  \item Nickolaos Tzeremes \email{bus9nt@econ.uth.gr} (\href{https://orcid.org/0000-0002-6938-3404}{ORCID})
}

}
\keyword{internal}
