% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.ardl}
\alias{predict.ardl}
\title{ARDL prediction}
\usage{
\method{predict}{ardl}(object, newdata, se.fit = FALSE, ...)
}
\arguments{
\item{object}{A fitted 'ardl' model of \code{\link[base]{class}} 'ardl',
created using \code{ardl}.}

\item{newdata}{A time series object (e.g., "ts", "zoo" or "zooreg") or a 
data frame containing the new values of \eqn{x} required to make 
predictions.}

\item{se.fit}{A switch indicating if standard errors are required.}
}
\value{
\code{predict} returns predictions of \eqn{y}.
}
\description{
Makes predictions based on fitted ARDL models, extending the 
\code{\link[stats]{predict}} generic function for the 
\code{\link[ARDL-package]}. It is designed to take advantage of lags of the 
dependent variable from the fitted model and new values of the independent
variable(s) to calculate predictions.
}
\section{Calculation}{

Details of the calculation, Kleanthis' matrices detail
}

\examples{
## Prediction example -----

# Indirectly from an ARDL
ardl_3132 <- ardl(LRM ~ LRY + IBO + IDE, data = denmark, order = c(3,1,3,2))
recm_3132 <- recm(ardl_3132, case = 2)
}
\seealso{
\code{\link{ardl}}
}
\author{
Kleanthis Natsiopoulos, \email{klnatsio@gmail.com}

Daniel Finnan, \email{dan@custom-made.org.uk}
}
\keyword{models}
\keyword{ts}
